fix typo
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@24863 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/src-java/x10/rtt/Types.java
@@ -694,7 +694,7 @@ public class Types {
         if (isNumericType(origRTT)) {
         	return ((java.lang.Number) typeParamOrAny).intValue();
         } else {
-        	if (typeParamOrAny instanceof x10.core.UInt) {return x10.core.UInt.$unbox((x10.core.Byte)typeParamOrAny);}
+        	if (typeParamOrAny instanceof x10.core.UInt) {return x10.core.UInt.$unbox((x10.core.UInt)typeParamOrAny);}
         	else if (typeParamOrAny instanceof java.lang.Integer) {return (java.lang.Integer)typeParamOrAny;}
         }
         throw new java.lang.ClassCastException("x10.lang.UInt");
